[Fix][Mamba] Add dtype parameter to DaCumsumFwd workload/tests
Make DaCumsumFwd consistent with other SSD ops by adding dtype parameter:
- Add dtype to DaCumsumFwdFixture and DaCumsumFwdTest
- Update reference implementation to return dt_out in target dtype
- Restore default dtype=float16 in Op/Kernel (now consistent with tests)

This resolves the design inconsistency where DaCumsumFwd had different
default behavior (float32) than its usage in Mamba2 (dtype). Now both
standalone and Mamba2 usage paths follow the same typed-dt contract.

+++ tileops/kernels/mamba/da_cumsum.py
@@ -235,7 +235,7 @@ class DaCumsumFwdKernel(Kernel):
         chunk_len: int,
         n_heads: int,
         seq_len: int,
-        dtype: torch.dtype = torch.float32,
+        dtype: torch.dtype = torch.float16,
         dt_softplus: bool = False,
         has_dt_bias: bool = False,
         dt_min: float = 0.0,
